Skip build steps if no packages were updated
In the build updated packages workflow, we may end up with a situation
where the list of updated packages is empty for a variety of reasons. If
this happens, there's no sense in trying to proceed with the build,
sign, rename, and upload steps. For these steps, we make them
conditional on the list of packages to build not being empty.