Use deterministic SOURCE_DATE_EPOCH for autobuilds
For automatic builds, use a deterministic value for SOURCE_DATE_EPOCH.
This should hopefully allow for package builds to be raced on different
build platforms and still, assuming the package is reproducible, result
in an identical package output.

It may be necessary to support installing packages that already exist in
my repository if it turns out that dependencies built with a different
SOURCE_DATE_EPOCH affect the resulting package. Another option would be
to set SOURCE_DATE_EPOCH to be the last commit timestamp for the
PKGBUILD for each package, but this is not possible with the current
design since we should not require Git to build packages that do not
need it.